+++ src/main/java/com/cells/cells/normal/compacting/ItemCompactingCellBase.java
@@ -123,7 +123,7 @@ public abstract class ItemCompactingCellBase extends Item implements IInternalCo
                     } else {
                         // Items stored but no compression found
                         tooltip.add("");
-                        tooltip.add("\u00a78" + I18n.format("tooltip.cells.compacting_cell.no_compression"));
+                        tooltip.add("\u00a7e" + I18n.format("tooltip.cells.compacting_cell.no_compression"));
                     }
                 }